Day 23, cutting edge compiler to conquer the task!
Fairly ad-hoc compiler for the small assembly language, targeting C.

The optimisations implemented are very much specific to the task at
hand, and are not generally correct (unless by accident). The code is
also very sloppy, and there *might* be some memory leaks, but hey, it
works™ for my inputs!